pyproject.toml
[project]
name = "whisper-typer"
version = "0.1.0"
description = "Speech-to-text dictation service with hotkey activation"
requires-python = ">=3.10"
dependencies = [
    "evdev>=1.7.0",
    "pyaudio>=0.2.14",
    "numpy>=1.24.0",
    "torch>=2.1.0",
    "transformers>=4.39.0",
    "accelerate>=0.26.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "PyGObject>=3.48.0",
    "rich>=13.0.0",
    "kokoro-onnx>=0.4.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
whisper-typer = "whisper_typer.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
